Guard against filtering by non-public states
Even though we don't present non-public facets to the user in the UI that's not to say they might not try monkeying with the URL to access them.  We introduce a guard to make sure that if the state param is not one of the public facets we redirect to the same url with `state=all` instead.

One wrinkle with this approach is that we add a method `public_petition_facets` to `ApplicationController` and make it a helper_method so we can use it in our controllers as well as in our views and helpers.  This helper method is *not* included in the `helper` object in our specs though, so we have to define the method to allow us to stub it and provide a canned response.