Make admin petitions filter also take facets from locale
Implementation is much like that for the public sphere, but we also have a list of names for how to describe the facets that are defined in the locale file too.  The dropdown takes the name from the locale if present, or humanizes the facet name if no name is present.