Add support to show pending, validated and sponsored petitions in admin
Have added two new facets since we want to expose all states for admin. The "collecting sponsors" facet combines sponsored and validated states, since these are to grouped for admin.
Have added the locale definitions for them. Have removed the selectable scope in the controller since it was restricting to moderated states only.