Add inspect method so that console doesn't load all records
If a blank search is made in the console then inspect falls back to the
relation via method missing and the default behaviour is to load it. This
could result in loading thousands of records so it's best that this isn't
accidentally caused by someone trying to fix something in an emergency.