Remove `AdminFormBuilder`
We were only using it in two remaining forms in admin and the HTML it generated was unlike the rest of the forms.  Our current style for forms does have a lot of boiler plate, so there is scope to extract a new FormBuilder once we know exactly how our HTML should be, but it's a bit in flux at the moment.