Remove admin tab_nav
It only has one element in it now: 'All petitions' and most activity will be from this page using different filters so we don't need a specific admin for it.  Users can get there form the home page hub, or just hacking the URL / changing the filters of the page we are (most likely) already on.

For: https://www.pivotaltracker.com/story/show/97990386