Use in_moderation instead of sponsored as a facet
This means that the admin hub link for "Awaiting moderation" now links through to a valid filtered list.  We also take the time to use existing scopes in our facet definitions if we can (e.g. in the `in_moderation` and `collecting_sponsors` facts we use the similarly named scopes in stead of re-writing `where(....)` clauses).