Fill in threshold explanation sections on home page
These blobs are always there regardless of the counts of petitions in those states (with a response and have been debated), unlike the counts at the top of the page.  However, we don't want to do the count queries twice so we rework `home_helper` to memoize the count decorator.  We also make the count decorator use a hash, not an array, and this lets us treat it differently in the two sections.  For the big counts at the top we just want to iterate and do the same thing each time, however for our new explanation sections we don't iterate as there's enough difference between the HTML in each, so we just use `[]` access to get the counts we want when we need them.