Don't set closed_at when publishing the petition
Since we have a fixed duration now we can determine the petitions to be
closed when we run our background job so presetting the closed_at when
publishing doesn't make sense and only complicates things. The quantity
of petitions to close on any given day will be approximately 40 based
on data from the old website so any performance impact from the
background job will be neglible.