Use id lookup in SignaturesController#signed
This page, the final step in signing a petition, had a url ``/petitions/<id>/signatures/<token>/signed`` which meant to find the signature we did a full table scan lookup on ``Signature#perishable_token``.  This is not particularly performant, particularly as ``perishable_token`` is not indexed, as such it was a bottleneck in the signing flow: these slow full table scans cause a lock on any writes to the table (e.g. for other signers at the initial signing or verification point in the flow).

The fix is to rework this action to be like ``SignaturesController#verify`` and ``SignaturesController#unsubscribe`` and lookup by id, but reject the request if the token does not match.  This means moving the url from ``/petitions/<id>/signatures/<token>/signed`` to ``/signatures/<id>/signed/<token>``.