Avoid extra queries in ``record_new_signature_for``
Change to `find_or_intialize_by` in `.for` and let `#record_new_signature_for` handle the `new_record?` vs. `persisted?` cases differently.  For persisted records we do what we always did (an atomic update and reload), for new records we save the instance with a signature count of 1.

This means we only do an INSERT for the "first signer of the petition from a constituency" rather than an INSERT (create the row), UPDATE (set the signature count), SELECT (reload the instance).