[restructure_indexes] Remove unnecessary indexes on the signatures table
The signatures table is going to be very large.

The state field is going to be a very small set of values, so the cardinality of
an index on just the single "state" field is going to be very poor.

I also removed the index on petition_id and state too. Queries on the table
always include an email field - and we never fetch items from the table just
using the petition and id alone.
Removing this means we can get away from maintaining a large and unnecessary
index.