Avoid reload in record_new_signature
To avoid doing a reload and issuing an extra SQL SELECT query during the validate! transaction for already saved instances we won't run reload. Instead we edit the instance in-place with a `signature_count + 1`.

We know this is a bad assumption because we could have two instances of constituency_petition_journal call record_new_signature close together.  This is why we do an atomic SQL update.  However, as the `#record_new_signature` is only used by `.record_new_signature_for` and the return value of that is ignored, it's not a huge problem.  As long as the DB is updated properly we can deal with the in-memory object being off.

We use `raw_write_attribute` to edit the in-memory instance without triggering any of the dirty tracking stuff so the object looks like it hasn't changed (or has been through a reload).