Move the creator_signature.validate! call out of transaction
It's most likely that we don't need to do this at all because the creator signature should be validated way earlier in the process, before any other signature even gets this far.  However, there's always a tiny chance so we leave it in.  However we'll take it out of the transaction (as it'll cause it's own transaction if it does indeed run).