Add ``Petition#deadline``
This allows us to only worry about a single method name to get the closing date for a petition.  For open petitions it asks the Site for the `closed_at_for_opening` providing its own `open_at` timestamp, for closed petitions it just returns its own `closed_at` timestamp.

`Site.closed_at_for_opening` is the mirror of the existing `Site.opened_at_for_closing`.  The original returns the opening time for a given closing time, and the new one returns the closing time for a given opening time.  Both use the petition_duration figure to go back or forwards in time.