Use `deadline` in views in preference to `closed_at`
Only where it makes sense, for example we don't need to use it in the _closed_petition_show partial as we explicitly know that these petitions have been closed and have a closed_at timestamp.  It's mostly in admin where we're not sure if the petition at hand will be a closed or open one that we need to use deadline.