Update data-generator.rake to create 'closed' petitions properly
Since e6e68f1c7a8cd1f8cd2967a5d27f2a38d157458e we have closed as a real state, not just a calculated one and don't always populate closed_at until the petition has actually been closed.  The data generator was forgotten and still created things with a `state` of 'open' and a `closed_at` in the past when it was creating a closed petition.

Not any more.