Make `Petition#deadline` cope with a missing open_at
Only petitions that have gone through `#publish!` will have an `open_at` so we need `#deadline` to cope with that.  Returning nil seems the most obvious as `closed_at` on these instances would've been nil anyway.

We simplify the logic to a single conditional too; although this does mean that we will use any closed_at that happens to be set on an open petition in favour of calculating it.  It shouldn't happen though so lets see.