Rework html and css for petition-meta
Again we namespace and prefix, but we also use a dl with dt+dd pairs.  The dl uses mostly the styling for an h2, and the dd uses mostly the styling of a p.  We only pull in the bits we need rather than extending our definitions though as that means we have the ability to change the styling independently.