Remove show page for petition action controllers
Now we use the petition_actions nav for displaying forms all the show templates were the same, so we might as well make that explicit and have each petition action controller render the admin/petitions/show template.

Note: PetitionDetailsController has a different show, so we leave it and ModerationController doesn't have a show at all (the form is inlined on normal show for todolist petitions to avoid an extra click).