Add some JS to show / hide rejection interface when moderating
There's no point seeing the rejection interface if you choose approve, and there's no point seeing it before you decide what to do.  We do always show it if there's an error on the object though.

The impl is just some inline JS because we already use inline JS for the rejection description stuff in the rejection interface.  Even if this isn't great, at least it's consistent.