Force early definition of attribute methods
Active Record will lazily define attribute methods when the model is
first instantiated but if it's come from the cache then that never
happens causing 'no superclass method' errors. By calling the method
`define_attribute_methods` directly we can trigger that definition
when the application boots and the class is eager loaded.