Add feature to describe moderators flagging a petition
On the moderation screen they should be able to choose "Flag" as an option as well as "Approve" or "Reject".  "Flagged" petitions are still considered as being in moderation, but have a state that can be identified separately from the existing pending/validated states.  The goal here is to identify that someone has looked at the petition, but don't yet know if it should be published or rejected.  This stops the situation where a moderator looks at a petition, can't decide to publish or reject and while they're seeking advice a different moderator goes to the moderation queue and moderates the same petition.