Give `fields_for` the object in reject form UI
Because ``Petition#rejection`` is not an ``accepts_nested_attributes_for``, when we use ``f.fields_for :rejection`` in the reject form UI it doesn't pull in the rejection object for us.  This means when we present an error, you lose anything you'd typed into the text area and have to reselect the code.

We also make sure to present the errors on the details, and change how we decide to hide the rejection UI in JS.  We do it via JS by looking for a .error-message element within the rejection controls UI.