Style admin petitions list for flagged petitions differently
We remove the petition_id class from the table as it wasn't used.  We also add some classes to identify this table as a list of petitions, and stop using the petition-action class that caused a visual bug because it was the same as one used for actions on the petition show page.