Support longer default email addresses in Site
We had a limit of 50 on email_from, but the new default (that has address and name) is 76.  We bump the limit to 100 and while we're at it change the default on the column to include the name as well as the address.

We also change the default in the db for the feedback email.