Add a feature to describe simple tag searching for admin
The admin users will be able to record a "tag" in the notes field using "[a tag]", and then search for those by entering "[a tag]" in the search box on the admin site.  The search will only look in the notes field for the petitions and return those with an exact match for the supplied tag.