Add Petition#tagged_with
This takes a single "tag" and searches the admin notes to find that tag.  Tags are wrapped in square brackets (e.g. `[foo]` is the tag `foo`) and this allows for spaces in the tag name(e.g. `[foo bar] is the tag `foo bar` not the tag `foo` and the tag `bar`).

We just use a SQL like query to do the match as we're filtering out any petitions that don't have notes so it shouldn't be too bad performance-wise, and it's not super clear how to do exact match phrase queries with postgres full text search.