Don't show "Flag" option for already flagged petitions
Once a petition is flagged you can't flag it anymore so it's a pointless option.  It's also a stronger signal to the user that the flag action they've taken has worked because we send them back to the show page and the UI has changed.  Before this commit all that changed on the page is the state in the sidebar, which they could easily might miss.