Implement request-based routing spec example group
The built-in RSpec routing spec example group is based upon the assertions
available in Rails. However these don't work in an integration/request
context so are essentially controller specs. This means that the full stack
isn't exercised so we need to implement our own routing spec example group
that does so we can test the constraints.