Prevent unnecessary SQL queries when rendering petitions/show
By checking first for the presence of the timestamp columns we can prevent
the execution of two SQL queries for most petitions as they never get
enough signatures for a government response or a parliamentary debate.

It could be argued that the use of a predicate method to return a value is
somewhat surprising but this is allowed by Ruby and is used elsewhere, e.g:

  $ irb
  >> 0.nonzero?
  => nil
  >> 5.nonzero?
  => 5

Since we will be caching the petitions/show content this probably won't save
much overall in the end but it's still better to be efficient as possible.