Add cache key class to build digests for caching
In an ideal world you'd be able to just call `cache` with a single argument
or at most a suffix to indicate which part of the template you're caching,
however our template structure is messy so we need to construct complicated
cache keys. This adds a class to extract this logic out of the view and into
a configuration file.