The assigns hash in a view uses string keys
There is some confusion over whether to use symbol vs. string keys when
testing assigns - Action View stores them with string keys but all the
examples for specs use symbol keys. This method of testing is deprecated in
Rails 5 as the intention is to make controller specs really request specs.