Fix race condition with sigining at moderation threshold
There existed a race condition where if multiple people confirmed their
sponsoring a petition at the same time the signature count could jump
over the threshold for moderation and the state not change from validated
to sponsored.

This fixes it by adding a timestamp column that records when the petition
reached the moderation threshold and then updating the state if the count
is at or above the threshold and the timestamp is nil. This is similar to
the approach taken for responses and debate outcomes.

There still is a possibility of multiple emails around the threshold level
but they were tweaked to show the signature count and would stop afterwards
due to the redirect to the moderation info page in the before action.