Clear form groups
When there's an error on just the citizenship checkbox Chrome thinks there
is space to put the submit button to the right of the form-group elements.
Fix this by clearing form groups as they're floated. The reasons that they
are floated are unclear (possibly something to do with multiple elements
on a single line) but rather than remove the float we chose to clear them
as it's possible that there maybe browser compatibility reasons that the
float on form groups is enabled.

https://www.pivotaltracker.com/story/show/98640970