Email sponsors when the petition is published
After a sponsor signs a petition we say on the thank you page that we
will email them when the petition is published but we currently don't.
This commit fixes that by sending individual emails to each of the
sponsors rather than by bcc since the original message to the creator
is personalized so isn't appropriate for a bcc.

https://www.pivotaltracker.com/story/show/98637502