Only change to a sponsored state if we're collecting
Add a guard clause to protect against changing the state to sponsored if
the moderation_threshold_reached_at column is nil but the petition has
been published. This is unlikely to be the case in production but there
is a small possiblity where a petition is published before reaching the
threshold for moderation.