Don't redirect from thank you page when at maximum sponsors
If we redirect away from the thank you page when a petition has the maximum
number of sponsors then the final sponsor sees the moderation page. This is
somewhat confusing to users so don't do this. The same rationale also applies
to the sponsored page - if a sponsor has had an email they should be able to
validated their email address.

https://www.pivotaltracker.com/story/show/98647724