Patch queue adapter for better reporting in AppSignal
Currently all the jobs are reported under the same name in AppSignal.
This is a dirty way to fix it but there is no other option as adding
a `appsignal_name` method to the job class doesn't work since that's
not the object that gets passed to AppSignal's Delayed Job plugin.