Replace '+' with '%20' in sharing links
Mailto links in some (all?) browsers don't decode '+' like they do for
HTTP urls so convert them to use percent-encoding. Unfortunately the
URI.escape method has been marked as obsolete for some time and there
is generally some confusion[1] over what is the right method to use.

Given this the simplest fix is to replace '+' with '%20' since we know
that `to_query` will continue to work in later versions of Rails.

[1]: http://stackoverflow.com/questions/2824126