Move model error messages to locale file
This tidies up the error messages and validations by moving the custom
messages out into the locale file and standardising on full error messages
on each field. Also tweaks the output to only show the first error message
on an attribute otherwise you get ugly messages like:

  Postcode must be completed Postcode not recognised

This generally works well - first they get prompted to fill in a value and
then an invalid message occurs if the value they type is incorrect.

https://www.pivotaltracker.com/story/show/95703388