Add 400 and 406 error pages
If someone types a link and then runs on with the sentence without a
space most autolinkers on Twitter, Facebook, etc. will add the first
word of the next sentence as the request format. To handle these
requests for an unknown format we can add a page for the 406 error code.

Also sometimes links are mangled and have invalid URL encoding so add
a page for the 400 error code.