Fix page title on local petitions without a postcode
If the user doesn't enter a postcode when searching from the home page
or they navigate to the page directly then the page title helper doesn't
have a count to pick which translation to use. For this case set the
count to -1 so that the :other key is used from the translation.