Fix memory leak when queuing jobs inside a job
Active Job uses a single log subscriber class[1] for both perform and
enqueue events. The consequence of this is that any queue events that
happen inside a perform event are treated as child events[2] and so get
added to the event object for the perform event.

This causes us a problem when queuing the individual emails inside a
background job as it means that even though we are using `find_each`
to batch the record loading we eventually end up with 100,000+ records
loaded into memory as part of the event object.

As a temporary fix we can split the log subscriber into two classes -
one for perform events and one for enqueue events that prevents the
enqueue events being treated as child events. It is however possible
that this behaviour is intentional and desired so we should start
considering other approaches for sending bulk email responses.

I've create rails/rails#21036 to track this issue.

[1]: https://github.com/rails/rails/blob/v4.2.3/activejob/lib/active_job/logging.rb#L53-L103
[2]: https://github.com/rails/rails/blob/v4.2.3/activesupport/lib/active_support/subscriber.rb#L88