Add job concerns to autoload paths
There is a common confusion between autoload paths and eager load paths
in Rails. Whilst all eager load paths are autoloaded it's not always true
that an autoload path is eagerly loaded. In the case of app/jobs/concerns
we want the path added to the autoload paths and we also want to mark it
as eagerly loaded - this is necessary so that Rails will find a top level
constant inside the concerns directory without having a Concerns namespace.