Fix Appsignal when used with multiple DJ workers
Similar to the fix for Puma we need to let Appsignal know that the process
has been forked so that it can re-establish connections. Unfortunately there
is no good way to do this so we have to override the after_fork class method
in an initializer to call Active Record's establish_connection as well as
the Appsignal forked! method.