Dynamically generate robots.txt
The standard robots.txt format is very limited in that it doesn't support
any kind of wildcard or pattern matching for entries and once a directory
is excluded you can't allow any entries in that directory. To workaround
this limitation we can dynamically generate a list of petition urls that
we want to exclude from search engines.