Add petition email crud feature
This adds the ability to edit, update and delete items of other
parliamentary business. For the email to send it needs to be queued
by pressing the 'Email {n} petitioners' at least once - multiple
presses don't cause multiple emails to be sent and editing an email
and pressing 'Save without emailing' won't prevent the sending of
a previously queued email. The only way to stop an email after it
has been sent is to delete it. This causes the job to error which
is rescued and ignored.