Reinstate journalling
This reverts the PR to temporarily disable them[1].  We're pretty confident that moving count updates outside the lock[2] has resolved the majority of the write contention we were seeing under heavy load so we can re-instate these counts.

Note that we will need to run a separate task to recalculate the counts as they are currently out of date.

[1]: https://github.com/alphagov/e-petitions/pull/413
[2]: https://github.com/alphagov/e-petitions/pull/414