Fix race condition when creating journals
A journal may have been created by another thread/process between the
find_or_initialize_by and the update_attributes calls causing a duplicate
record error. We've not seen this in production because the chance of it
occurring is low due to the number of constituencies (650) and the fact
that the country journal for the United Kingdom will have been set up
as part of the sponsor gathering process (other country journals have
very low signature counts).