Add support for queue priorities via named queues
Active Job in Rails 4.2 doesn't support setting a queue priority but does
allow setting a queue name. Whilst we could dedicated one worker to processing
the validation emails only it would give us more flexibility if the validation
emails took priority over the bulk emails for all workers.

To enable this capability this commit adds a before_create callback to the
Delayed Job Active Record model that sets the priority of the job based on a
predefined mapping of queue names to priorities.