Really fix race condition in journals
In 97b9839 the find_or_create_by was changed to find_or_initialize_by in
an attempt to save a query but unfortunately this introduced a race
condition which was supposed to be fixed by 1b07b99. However it wasn't
fixed properly because the id would be nil when the query was constructed.

This commit really fixes it at the cost of a possible extra query or two
in the case where another process creates it simultaneously.