[debate_outcome_image_105664380] Add support for the debate-outcome image
Updated image of the voting chamber so that it doesn't appear to always be
completely full. Administrators can now upload a representative image of
the chamber, which will be displayed instead of the placeholder.

It's possible to upload a slightly different image and have the image
automatically resized, but admins will probably get a better looking result if
they upload an image of 1260.0px by 710.0px exactly

The images are stored on S3 and proxied via the webserver at a /attachments/
url, so as to avoid needing separate SSL certificates. We also don't want
browsers to directly fetch files from S3.

Main Changes:

- Added the Paperclip gem to support storage of files on AWS.
- Added a field to DebateOutcome for the associated image.
- Added some validation of the image size and ratios.
- Added tests for the above.
- Included Fog AWS gem, since it's more flexible than the AWS-SDK gem.
  Additionally, Paperclip doesn't currently support AWS-SDK > 2, which
  is required for CodeDeploy, our AWS deployment system.

Other:

- Improved css styling of file upload field.
- Made some changes to the styling of the house of commons image when the
  file sizes are slightly different from expected