Retry signature validation if the schema changes
When a schema change occurs on the signatures table the prepared statement
cache for the lock becomes invalid and the PostgreSQL adapter will raise
PG::InFailedSqlTransaction. Because the transaction is already invalid any
attempt to deallocate the statement fails - see rails/rails#12330.

Whilst any general approach that retries the transaction block after
clearing the cache is not a workable solution it's one that can work
in this instance. We only retry the once - if clearing the cache doesn't
help then the cause of the transaction failure is somewhere else.