Fix migrations task for country petition journals
A `TRUNCATE` followed by an `INSERT ... FROM` doesn't really work that
well on a busy site like the petitions website because new journals will
start to appear before the `INSERT` completes. Fix this by doing it in
Ruby and locking the journals before calculating the signatures.

To do this in a performant manner we need an index on the petition_id
and location_code for the signatures table. We need to do this
concurrently otherwise it will block writes to the table, preventing
anyone from signing any petitions.