Use HTML parser to validate document markup
As the status of polyglot markup is either deprecated or undefined it's
better to use a HTML parser to validate the markup. The issue is that it
seems that Nokogiri doesn't support any HTML5 tags so we need to filter
out the errors created by those. Adding the tags to Nokogiri directly
appears to be non-trivial.