Override links_in_email method from email_spec gem
The links_in_email method uses URI.extract to pull out the links from
the default body part (in our case the HTML part). Unfortunately this
doubles the number of links since we are using the url as the link text.

Fix this by using our text_and_links_in_email method to get the url.
This method returns the link and text as an array of pairs so doesn't
have the same problem.