Move tokens from a path param to a query param
When we get more than 50,000+ signatures in a day Google Analytics
will send a 'Too many URLs' alert because the tokens are embedded
in the path, thus making each one a unique URL.

We can't tell GA to ignore parts of the URL but it can ignore named
query parameters[1], so by moving the token to a query parameter we
can ignore it and prevent any future alerts.

Fixes #448.

[1] https://support.google.com/analytics/answer/1010249?hl=en