Search for emails case-insensitively in the admin
The default behaviour of PostgreSQL is for varchar columns to be case
sensitive so to prevent people signing petitions again by altering the
case of their email address we force it to lowercase when we create
the signature record. Since this is not visible to the signer when they
raise a query with the moderation team they give their upper/lower
version of the email address and when it's copied and pasted into the
search box in the admin it returns no results because the search is
case sensitive.

One way to fix this would be to put a case-insensitive collation on the
email address column in the signatures table but given the number of
rows we have in the table it could mean taking the website down. So the
simplest fix is just to make the `for_email` method convert the email
address to lowercase before performing the search.

https://www.pivotaltracker.com/story/show/120103205