Sort moderation queue by the most recent threshold reached
Petitions in the moderation queue were being sorted by the most recently
created which meant that petitions that were created some time ago but
had for some reason take some time to gather enough sponsors ended up at
the bottom of the list.

It makes more sense to sort by the most recent ones to have passed the
moderation threshold first because petitions that have been waiting a
while for moderation are likely to have been flagged or contentious in
some way and are awaiting feedback before moderation takes place.

https://www.pivotaltracker.com/story/show/120102803