Don't write updates to the signature_count on every signing
Due to excessive write contention on the signature column the PostgreSQL locks
table is filling up. To work around this increment a memcached counter and then
write that back to the signature_count column every 5 minutes. Any discrepancies
should be picked up by the nightly PetitionCountJob that checks for any invalid
signature counts.

N.B. this means that the method cached_signature_count should be used when an
up-to-date value is required for display or use in an email.